[ref_phs, mpi] red phase est method1 mpi test

+++ tests/test_ref_phs_est.py
@@ -2,12 +2,10 @@ from __future__ import print_function
 import glob
 import os
 import shutil
-import subprocess
 import sys
 import tempfile
 import unittest
 import pytest
-from itertools import product
 import numpy as np
 
 from pyrate import config as cf
@@ -17,7 +15,7 @@ from pyrate.reference_phase_estimation import (estimate_ref_phase,
                                                ReferencePhaseError)
 from pyrate.scripts import run_prepifg
 from pyrate.scripts import run_pyrate
-from pyrate.shared import Ifg
+from pyrate import mpiops
 from tests import common
 from tests.common import SYD_TEST_DIR, sydney_data_setup
 
@@ -478,30 +476,35 @@ def modify_config(request, tempdir, get_config):
     shared.copytree(common.SYD_TEST_GAMMA, params_dict[cf.OBS_DIR])
     params_dict[cf.IFG_FILE_LIST] = os.path.join(
         params_dict[cf.OBS_DIR], 'ifms_17')
-    params_dict[cf.OUT_DIR] = tempdir()
-
     params_dict[cf.PARALLEL] = 0
     params_dict[cf.APS_CORRECTION] = 0
-
     yield params_dict
     # clean up
-    shutil.rmtree(params_dict[cf.OUT_DIR])
     shutil.rmtree(params_dict[cf.OBS_DIR])
 
 
 def test_refpixel_mpi(mpisync, tempdir, modify_config, ref_est_method):
 
     params_dict = modify_config
+    if mpiops.rank == 0:
+        outdir = tempdir()
+    else:
+        outdir = None
+    outdir = mpiops.comm.bcast(outdir, root=0)
+    params_dict[cf.OUT_DIR] = outdir
     params_dict[cf.REF_EST_METHOD] = ref_est_method
     xlks, ylks, crop = cf.transform_params(params_dict)
-    base_unw_paths = cf.original_ifg_paths(
-        params_dict[cf.IFG_FILE_LIST])
-    # dest_paths are tifs that have been geotif converted and multilooked
-    dest_paths = cf.get_dest_paths(
-        base_unw_paths, crop, params_dict, xlks)
 
-    # create the dest_paths files
-    run_prepifg.gamma_prepifg(base_unw_paths, params_dict)
+    base_unw_paths = cf.original_ifg_paths(params_dict[cf.IFG_FILE_LIST])
+    # dest_paths are tifs that have been geotif converted and multilooked
+    dest_paths = cf.get_dest_paths(base_unw_paths, crop, params_dict, xlks)
+
+    # run prepifg, create the dest_paths files
+    if mpiops.rank == 0:
+        run_prepifg.gamma_prepifg(base_unw_paths, params_dict)
+
+    mpiops.comm.barrier()
+
     refpx, refpy = run_pyrate.ref_pixel_calc(dest_paths, params_dict)
     run_pyrate.ref_phase_estimation_mpi(dest_paths, params_dict, refpx, refpy)
     ifgs_mpi = sydney_data_setup(datafiles=dest_paths)
@@ -511,21 +514,22 @@ def test_refpixel_mpi(mpisync, tempdir, modify_config, ref_est_method):
     params_dict_old = modify_config
     params_dict[cf.OUT_DIR] = tempdir()
     params_dict[cf.REF_EST_METHOD] = ref_est_method
-    xlks, ylks, crop = cf.transform_params(params_dict)
-    base_unw_paths = cf.original_ifg_paths(
-        params_dict[cf.IFG_FILE_LIST])
-    dest_paths = cf.get_dest_paths(
-        base_unw_paths, crop, params_dict, xlks)
-    run_prepifg.gamma_prepifg(base_unw_paths, params_dict_old)
-    ifgs = sydney_data_setup(datafiles=dest_paths)
-    ref_phs, ifgs = estimate_ref_phase(ifgs, params_dict, refpx, refpy)
-
-    for i, j in zip(ifgs, ifgs_mpi):
-        np.testing.assert_array_almost_equal(i.phase_data, j.phase_data,
-                                             decimal=2)
-        i.close()
-        j.close()
-    shutil.rmtree(ifgs_mpi_out_dir)  # remove mpi out dir
+    if mpiops.rank == 0:
+        xlks, ylks, crop = cf.transform_params(params_dict)
+        base_unw_paths = cf.original_ifg_paths(
+            params_dict[cf.IFG_FILE_LIST])
+        dest_paths = cf.get_dest_paths(
+            base_unw_paths, crop, params_dict, xlks)
+        run_prepifg.gamma_prepifg(base_unw_paths, params_dict_old)
+        ifgs = sydney_data_setup(datafiles=dest_paths)
+        ref_phs, ifgs = estimate_ref_phase(ifgs, params_dict, refpx, refpy)
+        for i, j in zip(ifgs, ifgs_mpi):
+            np.testing.assert_array_almost_equal(i.phase_data, j.phase_data,
+                                                 decimal=2)
+            i.close()
+            j.close()
+        shutil.rmtree(ifgs_mpi_out_dir)  # remove mpi out dir
+        shutil.rmtree(params_dict[cf.OUT_DIR])  # remove serial out dir
 
 
 if __name__ == '__main__':
